+++ MPGES.LK/Core/Models/InvoiceModel.swift
@@ -22,9 +22,9 @@ public struct InvoiceModel: Decodable {
     var year: Int
     var statusSaldoId: Int
     var statusSaldo: StatusSaldoModel
-    var saldo: Double
-    var debet: Double
-    var credit: Double
+    var saldo: Decimal
+    var debet: Decimal
+    var credit: Decimal
     var balanceEndOfPeriod: Double?
     var isClose: Bool
 }
