+++ MPGES.LK/Core/Models/ReceivedDataAddNewTemplateModel.swift
@@ -19,7 +19,7 @@ public struct ReceivedDataAddNewTemplateModel: Encodable, Decodable {
     let previousDate: String
     let previousReceivedData: Int
     var receivedData: Int?
-    let tariffValue: Double
+    let tariffValue: Decimal
     var date: String
     let deviceId: Int
     let razryad: Int
